+++ Model/DistributionFunction.cs
@@ -4,10 +4,10 @@ namespace RandomNumberGenerationAndModeling.Model
 {
     public abstract class DistributionFunction : IProbabilityDistribution
     {
-        public abstract float MathExpectation { get; }
-        public abstract float Variance { get; }
-        public abstract float StandardDeviation { get; }
-        public abstract float GetValue(float x);
+        public abstract double MathExpectation { get; }
+        public abstract double Variance { get; }
+        public abstract double StandardDeviation { get; }
+        public abstract double GetValue(double x);
         public abstract override string ToString();
     }
 }
