+++ ViewModel/BoxMullerSamplerDialogViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Windows;
 using RandomNumberGenerationAndModeling.Model;
 
@@ -7,8 +6,8 @@ namespace RandomNumberGenerationAndModeling.ViewModel
 {
     public class BoxMullerSamplerDialogViewModel : GeneratorDialogViewModel<BoxMullerSampler>
     {
-        public float Shift { get; set; }
-        public float Scale { get; set; }
+        public double Shift { get; set; }
+        public double Scale { get; set; }
 
         public BoxMullerSamplerDialogViewModel(BoxMullerSampler boxMullerSampler, Window dialogView) : base(boxMullerSampler, dialogView)
         {
